// src/payload-types.ts
export interface Category {
  id: string;
  name: string;
  slug: string;
  headline: string;
  description: string;
  posts?: {
    docs?: (string | Post)[];
    hasNextPage?: boolean;
    totalDocs?: number;
  };
  updatedAt: string;
  createdAt: string;
}
// ...
export interface CaseStudy {
  id: string;
  title: string;
  introContent: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  industry?: string | null;
  useCase?: string | null;
  partner?: (string | null) | Partner;
  featuredImage: string | Media;
  layout?:
    | (
        | Callout
        | Cta
        | CardGrid
        | CaseStudyCards
        | CaseStudiesHighlight
        | CaseStudyParallax
        | CodeFeature
        | Content
        | ContentGrid
        | FormBlock
        | HoverCards
        | HoverHighlights
        | LinkGrid
        | LogoGrid
        | MediaBlock
        | MediaContent
        | MediaContentAccordion
        | Pricing
        | ReusableContentBlock
        | Slider
        | Statement
        | StepsBlock
        | StickyHighlights
        | ExampleTabsBlock
      )[]
    | null;
  slug?: string | null;
  url?: string | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (string | null) | Media;
  };
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "partners".
 */
export interface Partner {
  id: string;
  name: string;
  website: string;
  email: string;
  slug: string;
  /**
   * Set to inactive to hide this partner from the directory.
   */
  agency_status?: ('active' | 'inactive') | null;
  hubspotID?: string | null;
  logo: string | Media;
  /**
   * This field is managed by the Featured Partners field in the Partner Program collection
   */
  featured?: boolean | null;
  topContributor?: boolean | null;
  content: {
    /**
     * 1600 x 800px recommended
     */
    bannerImage: string | Media;
    overview: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    services: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    idealProject: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    };
    caseStudy?: (string | null) | CaseStudy;
    /**
     * Contributions to Payload. Must be a valid GitHub issue, pull request, or discussion URL from a repo in the 'payloadcms' organization.
     */
    contributions?:
      | {
          type: 'discussion' | 'pr' | 'issue';
          repo: string;
          number: number;
          id?: string | null;
        }[]
      | null;
    projects?:
      | {
          year: number;
          name: string;
          link: string;
          id?: string | null;
        }[]
      | null;
  };
  city: string;
  regions: (string | Region)[];
  specialties: (string | Specialty)[];
  budgets: (string | Budget)[];
  industries: (string | Industry)[];
  social?:
    | {
        platform: 'linkedin' | 'twitter' | 'facebook' | 'instagram' | 'youtube' | 'github';
        url: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "regions".
 */
export interface Region {
  id: string;
  name: string;
  /**
   * Must contain only lowercase letters, numbers, hyphens, and underscores
   */
  value: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "specialties".
 */
export interface Specialty {
  id: string;
  name: string;
  /**
   * Must contain only lowercase letters, numbers, hyphens, and underscores
   */
  value: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "budgets".
 */
export interface Budget {
  id: string;
  name: string;
  /**
   * Must contain only lowercase letters, numbers, hyphens, and underscores
   */
  value: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "industries".
 */
export interface Industry {
  id: string;
  name: string;
  /**
   * Must contain only lowercase letters, numbers, hyphens, and underscores
   */
  value: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Footer {
  id: string;
  brandName?: string | null;
  brandAccent?: string | null;
  tagline?: string | null;
  columns?:
    | {
        label: string;
        navItems?:
          | {
              link: {
                type?: ('reference' | 'custom') | null;
                newTab?: boolean | null;
                reference?:
                  | ({
                      relationTo: 'pages';
                      value: string | Page;
                    } | null)
                  | ({
                      relationTo: 'posts';
                      value: string | Post;
                    } | null)
                  | ({
                      relationTo: 'case-studies';
                      value: string | CaseStudy;
                    } | null);
                url?: string | null;
                label: string;
                customId?: string | null;
              };
              id?: string | null;
            }[]
          | null;
        id?: string | null;
      }[]
    | null;
  copyright?: string | null;
  updatedAt?: string | null;
  createdAt?: string | null;
}
// ...
export interface GetStarted {
  id: string;
  heading?: string | null;
  tabs?:
    | {
        label: string;
        content: {
          root: {
            type: string;
            children: {
              type: any;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
        blockName?: string | null;
        blockType: 'richTextBlock';
      }[]
    | null;
  sidebar?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  sidebarLinks?:
    | {
        link: {
          type?: ('reference' | 'custom') | null;
          newTab?: boolean | null;
          reference?:
            | ({
                relationTo: 'pages';
                value: string | Page;
              } | null)
            | ({
                relationTo: 'posts';
                value: string | Post;
              } | null)
            | ({
                relationTo: 'case-studies';
                value: string | CaseStudy;
              } | null);
          url?: string | null;
          label: string;
          customId?: string | null;
        };
        id?: string | null;
      }[]
    | null;
  meta?: {
    title?: string | null;
    description?: string | null;
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (string | null) | Media;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "partner-program".
 */
export interface PartnerProgram {
  id: string;
  /**
   * Select the form that should be used for the contact form.
   */
  contactForm: string | Form;
  hero?: {
    richText?: {
      root: {
        type: string;
        children: {
          type: any;
          version: number;
          [k: string]: unknown;
        }[];
        direction: ('ltr' | 'rtl') | null;
        format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
        indent: number;
        version: number;
      };
      [k: string]: unknown;
    } | null;
    breadcrumbBarLinks?:
      | {
          link: {
            type?: ('reference' | 'custom') | null;
            newTab?: boolean | null;
            reference?:
              | ({
                  relationTo: 'pages';
                  value: string | Page;
                } | null)
              | ({
                  relationTo: 'posts';
                  value: string | Post;
                } | null)
              | ({
                  relationTo: 'case-studies';
                  value: string | CaseStudy;
                } | null);
            url?: string | null;
            label: string;
            customId?: string | null;
          };
          id?: string | null;
        }[]
      | null;
    heroLinks?:
      | {
          link: {
            type?: ('reference' | 'custom') | null;
            newTab?: boolean | null;
            reference?:
              | ({
                  relationTo: 'pages';
                  value: string | Page;
                } | null)
              | ({
                  relationTo: 'posts';
                  value: string | Post;
                } | null)
              | ({
                  relationTo: 'case-studies';
                  value: string | CaseStudy;
                } | null);
            url?: string | null;
            label: string;
            customId?: string | null;
          };
          id?: string | null;
        }[]
      | null;
  };
  featuredPartners: {
    description?: string | null;
    partners: (string | Partner)[];
  };
  contentBlocks?: {
    beforeDirectory?:
      | (
          | Callout
          | Cta
          | CardGrid
          | CaseStudyCards
          | CaseStudiesHighlight
          | CaseStudyParallax
          | CodeFeature
          | Content
          | ContentGrid
          | FormBlock
          | HoverCards
          | HoverHighlights
          | LinkGrid
          | LogoGrid
          | MediaBlock
          | MediaContent
          | MediaContentAccordion
          | Pricing
          | ReusableContentBlock
          | Slider
          | Statement
          | StepsBlock
          | StickyHighlights
          | ExampleTabsBlock
        )[]
      | null;
    afterDirectory?:
      | (
          | Callout
          | Cta
          | CardGrid
          | CaseStudyCards
          | CaseStudiesHighlight
          | CaseStudyParallax
          | CodeFeature
          | Content
          | ContentGrid
          | FormBlock
          | HoverCards
          | HoverHighlights
          | LinkGrid
          | LogoGrid
          | MediaBlock
          | MediaContent
          | MediaContentAccordion
          | Pricing
          | ReusableContentBlock
          | Slider
          | Statement
          | StepsBlock
          | StickyHighlights
          | ExampleTabsBlock
        )[]
      | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}